requirements.txt
Flask
Pillow
flask_login
